Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-order-4-tla-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected keyword 'import'

Harness Error (TIMEOUT), message = null

NOTRUN Test evaluation order of modules
